rename test_IPM_Class by  test_IPM  & updadate
